Mobile Item 95: New Competitions Entry Functionality added in MHS GOLF
- Create inteface for Add and Remove Members.
- Integrate 'ApplyEventEntryV2' web service to confirm entry.
- Update Booking Entry screen.
- Add user as first member when tap on 'Add Player' [In Process]

+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/web/models/competitionsentrynew/Team.java
@@ -24,7 +24,7 @@ public class Team {
     private List<Player> Players = null;
     @SerializedName("EntryFee")
     @Expose
-    private Integer EntryFee;
+    private Double EntryFee;
 
     public Integer getZoneId() {
         return ZoneId;
@@ -66,11 +66,11 @@ public class Team {
         this.Players = Players;
     }
 
-    public Integer getEntryFee() {
+    public Double getEntryFee() {
         return EntryFee;
     }
 
-    public void setEntryFee(Integer EntryFee) {
+    public void setEntryFee(Double EntryFee) {
         this.EntryFee = EntryFee;
     }
